+++ python/GafferImageTest/__init__.py
@@ -111,6 +111,7 @@ from .CopyViewsTest import CopyViewsTest
 from .AnaglyphTest import AnaglyphTest
 from .LookTransformTest import LookTransformTest
 from .OpenColorIOAlgoTest import OpenColorIOAlgoTest
+from .OpenColorIOContextTest import OpenColorIOContextTest
 
 if __name__ == "__main__":
 	import unittest